refactor: 채팅 기능 다시 완전 분리
