+++ src/chat/types/common.ts
